% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBinManConsistency.R
\name{checkBinManConsistency}
\alias{checkBinManConsistency}
\title{Check \code{bin} and \code{man} directory consistency}
\usage{
checkBinManConsistency()
}
\description{
Check \code{bin} and \code{man} directory consistency
}
\details{
Check that all scripts in \code{bin} and \code{sbin} directories have corresponding
documentation in \code{man/man1}.
}
\note{
Updated 2021-03-01.
}
\examples{
## > checkBinManConsistency()
}
