% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInstalled.R
\name{checkInstalled}
\alias{checkInstalled}
\title{Check if program is installed}
\usage{
checkInstalled(name, required = TRUE, path = FALSE)
}
\arguments{
\item{name}{\code{character}.
Program name.}

\item{required}{\code{logical(1)}.
Is the program required or optional?}

\item{path}{\code{logical(1)}.
Display the path to program.
Calls \code{Sys.which()} internally.}
}
\value{
\code{logical}.
}
\description{
Check if program is installed
}
\note{
Updated 2020-08-11.
}
\examples{
checkInstalled("bash")
}
