% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSystem.R
\name{checkSystem}
\alias{checkSystem}
\title{Check system}
\usage{
checkSystem()
}
\description{
Check system
}
\details{
If you see this error, reinstall ruby, rbenv, and emacs:
}
\note{
Updated 2020-10-26.
}
\section{Ignoring commonmarker-0.17.13 because its extensions are not built.}{
Try: gem pristine commonmarker --version 0.17.13
}

\examples{
## > checkSystem()
}
