% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVersion.R
\name{checkVersion}
\alias{checkVersion}
\alias{checkMacOSAppVersion}
\alias{checkHomebrewCaskVersion}
\title{Check version}
\usage{
checkVersion(
  name,
  nameFancy = NULL,
  current = currentVersion(name),
  expected = expectedVersion(name),
  whichName,
  op = c("==", ">="),
  required = TRUE
)

checkMacOSAppVersion(name)

checkHomebrewCaskVersion(name)
}
\arguments{
\item{name}{\code{character(1)}.
Program name.
\code{current} and \code{expected} arguments use this value for version checks when
they are left unset.}

\item{nameFancy}{\code{character(1)} or \code{NULL}..
Fancy name for use in check messages.
Inherits \code{name} when left \code{NULL}.}

\item{current, expected}{\code{character(1)}.
Current or expected version number.}

\item{whichName}{\code{character(1)}, \code{missing}, or \code{NULL}.
Program name to check in \code{PATH}.
Inherits from \code{name} when left unset.
Set \code{NULL} to skip any additional \code{PATH} checks.}

\item{op}{\code{character(1)}.
Operator to use for evaluation.}

\item{required}{\code{logical(1)}.
Is the program required or optional?}
}
\value{
Invisible \code{logical}.
}
\description{
Check version
}
\section{Functions}{
\itemize{
\item \code{checkMacOSAppVersion}: Check macOS app version.

\item \code{checkHomebrewCaskVersion}: Check Homebrew Cask version.
}}

\note{
Updated 2020-08-11.
}
\examples{
checkVersion("tmux")
}
