% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentVersion.R
\name{currentVersion}
\alias{currentVersion}
\alias{currentHomebrewCaskVersion}
\alias{currentMacOSAppVersion}
\alias{currentMajorVersion}
\alias{currentMinorVersion}
\title{Current version of installed program}
\usage{
currentVersion(x, .fun = "get-version")

currentHomebrewCaskVersion(x)

currentMacOSAppVersion(x)

currentMajorVersion(x)

currentMinorVersion(x)
}
\arguments{
\item{x}{\code{character}.
Program names.}

\item{.fun}{\code{character(1)}.
Koopa version function to call internally via \code{shell()}.
Intended for use with wrapper functions only.}
}
\value{
\code{character}.
Version character string.
}
\description{
Current version of installed program
}
\section{Functions}{
\itemize{
\item \code{currentVersion}: Current system command version.

\item \code{currentHomebrewCaskVersion}: Current Homebrew Cask version.

\item \code{currentMacOSAppVersion}: Current macOS app version.

\item \code{currentMajorVersion}: Current major version.

\item \code{currentMinorVersion}: Current minor version.
}}

\note{
Updated 2020-08-11.
}
\examples{
apps <- c("vim", "tmux")
if (goalie::allAreSystemCommands(apps)) {
    currentVersion(apps)
    currentMajorVersion(apps)
    currentMinorVersion(apps)
}

if (goalie::isMacOS()) {
    apps <- c("BBEdit", "RStudio")
    currentMacOSAppVersion(apps)
    casks <- c("bbedit", "rstudio")
    currentHomebrewCaskVersion(casks)
}
}
