% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerPruneStaleTags.R
\name{dockerPruneStaleTags}
\alias{dockerPruneStaleTags}
\title{Prune (delete) stale tags from DockerHub for a specific repo}
\usage{
dockerPruneStaleTags()
}
\description{
Prune (delete) stale tags from DockerHub for a specific repo
}
\note{
Updated 2021-03-02.
}
\examples{
## > dockerPruneTags()
}
