% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGenome.R
\name{downloadGenome}
\alias{downloadGenome}
\alias{downloadEnsemblGenome}
\title{Download a genome}
\usage{
downloadEnsemblGenome()
}
\description{
Download a genome
}
\note{
Updated 2021-01-04.
}
\examples{
## > downloadEnsemblGenome()
}
