% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drat.R
\name{drat}
\alias{drat}
\title{Build a package and commit to drat repo}
\usage{
drat()
}
\description{
Build a package and commit to drat repo
}
\note{
Updated 2021-01-04.
}
\examples{
## > drat("~/monorepo/r-packages/basejump")
}
