% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedVersion.R
\name{expectedVersion}
\alias{expectedVersion}
\alias{expectedHomebrewCaskVersion}
\alias{expectedMacOSAppVersion}
\alias{expectedMajorVersion}
\alias{expectedMinorVersion}
\title{Expected version}
\usage{
expectedVersion(x)

expectedHomebrewCaskVersion(x)

expectedMacOSAppVersion(x)

expectedMajorVersion(x)

expectedMinorVersion(x)
}
\description{
Expected version
}
\section{Functions}{
\itemize{
\item \code{expectedVersion}: Expected system command version.

\item \code{expectedHomebrewCaskVersion}: Expected Homebrew Cask version.

\item \code{expectedMacOSAppVersion}: Expected macOS app version.

\item \code{expectedMajorVersion}: Expected major version.

\item \code{expectedMinorVersion}: Expected minor version.
}}

\note{
Updated 2020-08-11.
}
\examples{
apps <- c("vim", "tmux")
expectedVersion(apps)
expectedMajorVersion(apps)
expectedMinorVersion(apps)
}
