% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{gitDefaultBranch}
\alias{gitDefaultBranch}
\title{Default Git branch}
\usage{
gitDefaultBranch(repo, remote = "origin")
}
\arguments{
\item{repo}{\code{character(1)}.
Git repository directory path.}

\item{remote}{\code{character(1)}.
Remote name.}
}
\description{
Default Git branch
}
\note{
Updated 2021-04-23.
}
\examples{
## > repo <- file.path("~", "git", "monorepo")
## > gitDefaultBranch(repo)
}
\seealso{
\itemize{
\item \code{koopa::git_default_branch} (shell).
}
}
