% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-header.R
\name{header}
\alias{header}
\alias{h1}
\alias{h2}
\alias{h3}
\alias{h4}
\alias{h5}
\alias{h6}
\alias{h7}
\title{Header level}
\usage{
h1(x)

h2(x)

h3(x)

h4(x)

h5(x)

h6(x)

h7(x)
}
\arguments{
\item{`character`.}{Text to display as header.}
}
\description{
Header level
}
\note{
Updated 2020-08-11.
}
\examples{
h1("Level 1")
h2("Level 2")
}
