% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAcidverse.R
\name{installAcidverse}
\alias{installAcidverse}
\title{Install Acid Genomics packages}
\usage{
installAcidverse()
}
\value{
Status from \code{install()}.
}
\description{
Install Acid Genomics packages
}
\note{
Updated 2020-10-28.
}
\examples{
## > installAcidverse()
}
