% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installRPackages.R
\name{installRPackages}
\alias{installRPackages}
\title{Install R packages}
\usage{
installRPackages()
}
\description{
Install R packages
}
\note{
Updated 2021-01-04.
}
\examples{
## > installRPackages()
}
