% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean-isCellar.R
\name{isCellar}
\alias{isCellar}
\title{Is a system command cellarized?}
\usage{
isCellar(x, pattern = "/cellar/")
}
\arguments{
\item{x}{\code{character}.
Application name.}

\item{pattern}{\code{character(1)}.
Case insensitive file path pattern matching string.}
}
\description{
Is a system command cellarized?
}
\note{
Updated 2020-08-11.
}
\examples{
isCellar(c("vim", "tmux"))
}
