% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{koopa-package}
\title{koopa}
\description{
Shell bootloader for bioinformatics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://koopa.acidgenomics.com/
    https://github.com/acidgenomics/koopa/}
  \item Report bugs at \url{https://github.com/acidgenomics/koopa/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
