% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPrograms.R
\name{listPrograms}
\alias{listPrograms}
\title{List user-accessible programs exported in PATH}
\usage{
listPrograms()
}
\description{
List user-accessible programs exported in PATH
}
\note{
Updated 2020-12-31.
}
\examples{
## > listPrograms()
}
