% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdownDeployToAWS.R
\name{pkgdownDeployToAWS}
\alias{pkgdownDeployToAWS}
\title{Deploy a pkgdown website to AWS S3 bucket}
\usage{
pkgdownDeployToAWS()
}
\value{
\code{logical(1)}.
}
\description{
Deploy a pkgdown website to AWS S3 bucket
}
\note{
Updated 2021-03-01.
}
\examples{
## > pkgdownDeployToAWS()
}
