% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntactic.R
\name{syntactic}
\alias{syntactic}
\alias{camelCase}
\alias{kebabCase}
\alias{snakeCase}
\title{Syntactic naming functions}
\usage{
camelCase()

kebabCase()

snakeCase()
}
\description{
Syntactic naming functions
}
\note{
Updated 2021-02-06.
}
