% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-ul.R
\name{ul}
\alias{ul}
\title{Unordered list}
\usage{
ul(x)
}
\arguments{
\item{x}{\code{character(1)}.
Text to be displayed as list items.}
}
\description{
Unordered list
}
\note{
Updated 2020-08-11.
}
\examples{
ul(c("Item 1", "Item 2"))
}
