% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlinkApp.R
\name{unlinkApp}
\alias{unlinkApp}
\title{Unlink an application}
\usage{
unlinkApp()
}
\arguments{
\item{x}{\code{character.}
Application names.}
}
\description{
Unlink an application
}
\note{
Updated 2021-01-04.
}
\examples{
## > unlinkApp(c("vim", "tmux"))
}
