% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlinkCellar.R
\name{unlinkCellar}
\alias{unlinkCellar}
\title{Unlink cellar files}
\usage{
unlinkCellar(apps)
}
\arguments{
\item{apps}{\code{character.}
Application names.}
}
\description{
Unlink cellar files
}
\note{
Updated 2020-08-09.
}
\examples{
## > unlinkCellar(c("vim", "tmux"))
}
