% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVersion.R
\name{checkVersion}
\alias{checkVersion}
\alias{checkGNUVersion}
\alias{checkMacOSAppVersion}
\alias{checkHomebrewCaskVersion}
\title{Check version}
\usage{
checkVersion(
  name,
  nameFancy = capitalize(gsub("-", " ", name)),
  current = currentVersion(name),
  expected = expectedVersion(name),
  op = c("==", ">="),
  required = TRUE
)

checkGNUVersion(name, ...)

checkMacOSAppVersion(name, ...)

checkHomebrewCaskVersion(name, ...)
}
\arguments{
\item{name}{\code{character(1)}.
Program name.
\code{current} and \code{expected} arguments use this value for version checks when
they are left unset.}

\item{nameFancy}{\code{character(1)} or \code{NULL}..
Fancy name for use in check messages.}

\item{current, expected}{\code{character(1)}.
Current or expected version number.}

\item{op}{\code{character(1)}.
Operator to use for evaluation.}

\item{required}{\code{logical(1)}.
Is the program required or optional?}

\item{...}{Passthrough to \code{checkVersion()}.}
}
\value{
Invisible \code{logical}.
}
\description{
Check version
}
\section{Functions}{
\itemize{
\item \code{checkGNUVersion}: Check GNU package version.

\item \code{checkMacOSAppVersion}: Check macOS app version.

\item \code{checkHomebrewCaskVersion}: Check Homebrew Cask version.
}}

\note{
Updated 2020-08-11.
}
\examples{
checkVersion("tmux")
}
