% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVersion.R
\name{checkVersion}
\alias{checkVersion}
\alias{checkMacOSAppVersion}
\alias{checkHomebrewCaskVersion}
\title{Check version}
\usage{
checkVersion(
  name,
  nameFancy = NULL,
  current = currentVersion(name),
  expected = expectedVersion(name),
  whichName,
  eval = c("==", ">="),
  required = TRUE
)

checkMacOSAppVersion(name)

checkHomebrewCaskVersion(name)
}
\description{
Check version
}
\section{Functions}{
\itemize{
\item \code{checkMacOSAppVersion}: Check macOS app version.

\item \code{checkHomebrewCaskVersion}: Check Homebrew Cask version.
}}

\note{
Updated 2020-08-09.
}
\examples{
checkVersion("tmux")
}
