% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker}
\alias{docker}
\alias{dockerBuildAllTags}
\title{Manage Docker images}
\usage{
dockerBuildAllTags()
}
\description{
Manage Docker images
}
\section{Functions}{
\itemize{
\item \code{dockerBuildAllTags}: Build all tags for a specific image.
}}

\note{
Updated 2021-01-04.
}
\examples{
## > dockerBuildAllTags()
}
