% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGenome.R
\name{downloadGenome}
\alias{downloadGenome}
\alias{downloadEnsemblGenome}
\alias{downloadGencodeGenome}
\alias{downloadRefSeqGenome}
\alias{downloadUCSCGenome}
\title{Download a genome}
\usage{
downloadEnsemblGenome()

downloadGencodeGenome()

downloadRefSeqGenome()

downloadUCSCGenome()
}
\description{
Download a genome
}
\note{
Updated 2021-01-31.
}
\examples{
## Ensembl.
## > downloadEnsemblGenome()

## GENCODE.
## > downloadGencodeGenome()

## RefSeq.
## > downloadRefSeqGenome()

## UCSC.
## > downloadUCSCGenome()
}
