% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drat.R
\name{drat}
\alias{drat}
\title{Build a package and commit to drat repo}
\usage{
drat()
}
\value{
\code{logical(1)}.
}
\description{
Build a package and commit to drat repo
}
\note{
Updated 2021-01-06.
}
\examples{
## > drat()
}
