% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAndMoveInSequence.R
\name{findAndMoveInSequence}
\alias{findAndMoveInSequence}
\title{Find and move files in sequence}
\usage{
findAndMoveInSequence()
}
\description{
Find and move files in sequence
}
\note{
Updated 2021-03-01.
}
\examples{
## > findAndMoveInSequence()
}
