% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAndMoveInSequence.R
\name{findAndMoveInSequence}
\alias{findAndMoveInSequence}
\title{Find and move files in sequence}
\usage{
findAndMoveInSequence(sourceDir, targetDir)
}
\description{
Find and move files in sequence
}
\note{
Updated 2020-08-11.
}
\examples{
## > findAndMoveInSequence(sourceDir, targetDir)
}
