% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{shell}
\alias{install}
\alias{uninstall}
\alias{updatePackages}
\alias{assert}
\alias{isCleanSystemLibrary}
\alias{camelCase}
\alias{kebabCase}
\alias{snakeCase}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{acidbase}{\code{\link[acidbase]{shell}}}

  \item{bb8}{\code{\link[bb8]{install}}, \code{\link[bb8]{uninstall}}, \code{\link[bb8]{updatePackages}}}

  \item{goalie}{\code{\link[goalie]{assert}}, \code{\link[goalie:check-scalar-isCleanSystemLibrary]{isCleanSystemLibrary}}}

  \item{syntactic}{\code{\link[syntactic]{camelCase}}, \code{\link[syntactic]{kebabCase}}, \code{\link[syntactic]{snakeCase}}}
}}

