% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{initDir}
\alias{realpath}
\alias{shell}
\alias{install}
\alias{installAcidverse}
\alias{installDefaultPackages}
\alias{uninstall}
\alias{updatePackages}
\alias{assert}
\alias{hasLength}
\alias{isADir}
\alias{isAFile}
\alias{isCleanSystemLibrary}
\alias{isSubset}
\alias{camelCase}
\alias{kebabCase}
\alias{snakeCase}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidBase}{\code{\link[AcidBase]{initDir}}, \code{\link[AcidBase]{realpath}}, \code{\link[AcidBase]{shell}}}

  \item{bb8}{\code{\link[bb8]{install}}, \code{\link[bb8]{installAcidverse}}, \code{\link[bb8]{installDefaultPackages}}, \code{\link[bb8]{uninstall}}, \code{\link[bb8]{updatePackages}}}

  \item{goalie}{\code{\link[goalie]{assert}}, \code{\link[goalie:check-scalar-hasLength]{hasLength}}, \code{\link[goalie:check-vector-isDirectory]{isADir}}, \code{\link[goalie:check-vector-isFile]{isAFile}}, \code{\link[goalie:check-scalar-isCleanSystemLibrary]{isCleanSystemLibrary}}, \code{\link[goalie:check-scalar-sets]{isSubset}}}

  \item{syntactic}{\code{\link[syntactic]{camelCase}}, \code{\link[syntactic]{kebabCase}}, \code{\link[syntactic]{snakeCase}}}
}}

